+++ grpc.gyp
@@ -972,10 +972,10 @@
         'src/core/lib/surface/wait_for_cq_end_op.cc',
         'src/core/lib/transport/batch_builder.cc',
         'src/core/lib/transport/bdp_estimator.cc',
-        'src/core/lib/transport/call_filters.cc',
+        'src/core/call/call_filters.cc',
         'src/core/lib/transport/call_final_info.cc',
         'src/core/lib/transport/call_size_estimator.cc',
-        'src/core/lib/transport/call_spine.cc',
+        'src/core/call/call_spine.cc',
         'src/core/lib/transport/connectivity_state.cc',
         'src/core/lib/transport/error_utils.cc',
         'src/core/lib/transport/handshaker.cc',
@@ -1476,10 +1476,10 @@
         'src/core/lib/surface/wait_for_cq_end_op.cc',
         'src/core/lib/transport/batch_builder.cc',
         'src/core/lib/transport/bdp_estimator.cc',
-        'src/core/lib/transport/call_filters.cc',
+        'src/core/call/call_filters.cc',
         'src/core/lib/transport/call_final_info.cc',
         'src/core/lib/transport/call_size_estimator.cc',
-        'src/core/lib/transport/call_spine.cc',
+        'src/core/call/call_spine.cc',
         'src/core/lib/transport/connectivity_state.cc',
         'src/core/lib/transport/error_utils.cc',
         'src/core/lib/transport/handshaker.cc',
@@ -2279,10 +2279,10 @@
         'src/core/lib/surface/version.cc',
         'src/core/lib/surface/wait_for_cq_end_op.cc',
         'src/core/lib/transport/batch_builder.cc',
-        'src/core/lib/transport/call_filters.cc',
+        'src/core/call/call_filters.cc',
         'src/core/lib/transport/call_final_info.cc',
         'src/core/lib/transport/call_size_estimator.cc',
-        'src/core/lib/transport/call_spine.cc',
+        'src/core/call/call_spine.cc',
         'src/core/lib/transport/connectivity_state.cc',
         'src/core/lib/transport/error_utils.cc',
         'src/core/lib/transport/handshaker.cc',
