+++ src/core/call/call_destination.h
@@ -17,8 +17,8 @@
 
 #include <grpc/support/port_platform.h>
 
+#include "src/core/call/call_spine.h"
 #include "src/core/lib/gprpp/orphanable.h"
-#include "src/core/lib/transport/call_spine.h"
 
 namespace grpc_core {
