+++ src/core/client_channel/client_channel_internal.h
@@ -24,10 +24,10 @@
 #include "absl/functional/any_invocable.h"
 #include "absl/log/check.h"
 
+#include "src/core/call/call_destination.h"
 #include "src/core/lib/gprpp/down_cast.h"
 #include "src/core/lib/gprpp/unique_type_name.h"
 #include "src/core/lib/resource_quota/arena.h"
-#include "src/core/lib/transport/call_destination.h"
 #include "src/core/load_balancing/lb_policy.h"
 #include "src/core/service_config/service_config_call_data.h"
 #include "src/core/telemetry/call_tracer.h"
