+++ src/core/client_channel/load_balanced_call_destination.h
@@ -17,9 +17,9 @@
 
 #include "absl/functional/any_invocable.h"
 
+#include "src/core/call/call_destination.h"
 #include "src/core/client_channel/client_channel.h"
 #include "src/core/lib/promise/context.h"
-#include "src/core/lib/transport/call_destination.h"
 #include "src/core/load_balancing/lb_policy.h"
 
 namespace grpc_core {
