+++ src/core/lib/surface/channel_init.h
@@ -32,13 +32,13 @@
 
 #include <grpc/support/port_platform.h>
 
+#include "src/core/call/call_filters.h"
 #include "src/core/lib/channel/channel_args.h"
 #include "src/core/lib/channel/channel_fwd.h"
 #include "src/core/lib/channel/channel_stack_builder.h"
 #include "src/core/lib/gprpp/debug_location.h"
 #include "src/core/lib/gprpp/unique_type_name.h"
 #include "src/core/lib/surface/channel_stack_type.h"
-#include "src/core/lib/transport/call_filters.h"
 #include "src/core/lib/transport/interception_chain.h"
 
 /// This module provides a way for plugins (and the grpc core library itself)
