+++ src/core/lib/transport/interception_chain.h
@@ -20,10 +20,10 @@
 
 #include <grpc/support/port_platform.h>
 
+#include "src/core/call/call_destination.h"
+#include "src/core/call/call_filters.h"
+#include "src/core/call/call_spine.h"
 #include "src/core/lib/gprpp/ref_counted.h"
-#include "src/core/lib/transport/call_destination.h"
-#include "src/core/lib/transport/call_filters.h"
-#include "src/core/lib/transport/call_spine.h"
 #include "src/core/lib/transport/metadata.h"
 
 namespace grpc_core {
