+++ test/core/transport/call_filters_test.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "src/core/lib/transport/call_filters.h"
+#include "src/core/call/call_filters.h"
 
 #include <vector>
