+++ test/core/transport/call_spine_benchmarks.h
@@ -19,6 +19,7 @@
 
 #include "benchmark/benchmark.h"
 
+#include "src/core/call/call_spine.h"
 #include "src/core/lib/event_engine/default_event_engine.h"
 #include "src/core/lib/event_engine/event_engine_context.h"
 #include "src/core/lib/gprpp/notification.h"
@@ -26,7 +27,6 @@
 #include "src/core/lib/promise/all_ok.h"
 #include "src/core/lib/promise/map.h"
 #include "src/core/lib/resource_quota/resource_quota.h"
-#include "src/core/lib/transport/call_spine.h"
 #include "src/core/lib/transport/transport.h"
 
 namespace grpc_core {
